+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## Current (in progress)
 
+- Add a `burnthemall` command [#25](https://github.com/etalab/cada.data.gouv.fr/pull/25)
+- Use `tqdm` for major commands with loops [#25](https://github.com/etalab/cada.data.gouv.fr/pull/25)
 - Python 3 [#28](https://github.com/etalab/cada.data.gouv.fr/pull/28)
 - Elastic Search 7.2 [#29](https://github.com/etalab/cada.data.gouv.fr/pull/29)
 - Mongo 4.1 [#29](https://github.com/etalab/cada.data.gouv.fr/pull/29)